Conflicts:
	src/FSAL/FSAL_LUSTRE/fsal_internal.h
	src/FSAL/FSAL_POSIX/Makefile.am
	src/FSAL/FSAL_POSIX/fsal_convert.c
	src/FSAL/FSAL_PROXY/Makefile.am
	src/FSAL/FSAL_SNMP/Makefile.am
	src/FSAL/FSAL_SNMP/fsal_attrs.c
	src/FSAL/FSAL_SNMP/fsal_compat.c
	src/FSAL/FSAL_SNMP/fsal_context.c
	src/FSAL/FSAL_SNMP/fsal_init.c
	src/FSAL/FSAL_SNMP/fsal_internal.c
	src/FSAL/FSAL_SNMP/fsal_internal.h
	src/FSAL/FSAL_SNMP/fsal_tools.c
	src/FSAL/FSAL_SNMP/fsal_xattrs.c
	src/FSAL/Makefile.am
	src/FSAL/access_check.c
	src/FSAL/common_methods.c
	src/include/FSAL/access_check.h
	src/include/FSAL/common_methods.h
	src/shell/Makefile.am

+++ src/Cache_inode/cache_inode_rename.c
@@ -84,17 +84,17 @@ cache_inode_status_t cache_inode_rename_cached_dirent(cache_entry_t * pentry_par
   *pstatus = CACHE_INODE_SUCCESS;
 
   /* Sanity check */
-  if(pentry_parent->internal_md.type != DIR_BEGINNING &&
-     pentry_parent->internal_md.type != DIR_CONTINUE)
+  if(pentry_parent->internal_md.type != DIRECTORY)
     {
       *pstatus = CACHE_INODE_BAD_TYPE;
       return *pstatus;
     }
 
-  /* BUGAZOMEU: Ne pas oublier de jarter un dir_cont dont toutes les entrees sont inactives */
+  /* BUGAZOMEU: Ne pas oublier de jarter un dir dont toutes les entrees sont inactives */
   if((removed_pentry = cache_inode_operate_cached_dirent(pentry_parent,
                                                          oldname,
                                                          newname,
+                                                         pclient,
                                                          CACHE_INODE_DIRENT_OP_RENAME,
                                                          pstatus)) == NULL)
     return *pstatus;
@@ -137,6 +137,7 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
 {
   cache_inode_status_t status;
   fsal_status_t fsal_status;
+  cache_inode_dir_entry_t *new_dir_entry;
   cache_entry_t *pentry_lookup_src = NULL;
   cache_entry_t *pentry_lookup_dest = NULL;
   fsal_attrib_list_t attrlookup;
@@ -153,10 +154,8 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
   pclient->stat.func_stats.nb_call[CACHE_INODE_RENAME] += 1;
 
   /* Are we working on directories ? */
-  if((pentry_dirsrc->internal_md.type != DIR_BEGINNING
-      && pentry_dirsrc->internal_md.type != DIR_CONTINUE)
-     || (pentry_dirdest->internal_md.type != DIR_BEGINNING
-         && pentry_dirdest->internal_md.type != DIR_CONTINUE))
+  if((pentry_dirsrc->internal_md.type != DIRECTORY)
+     || (pentry_dirdest->internal_md.type != DIRECTORY))
     {
       /* Bad type .... */
       *pstatus = CACHE_INODE_BAD_TYPE;
@@ -192,6 +191,7 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
   /* Check for object existence in source directory */
   if((pentry_lookup_src = cache_inode_lookup_no_mutex(pentry_dirsrc,
                                                       poldname,
+                                                      CACHE_INODE_JOKER_POLICY,
                                                       &attrlookup,
                                                       ht,
                                                       pclient,
@@ -224,6 +224,7 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
   /* Check if an object with the new name exists in the destination directory */
   if((pentry_lookup_dest = cache_inode_lookup_no_mutex(pentry_dirdest,
                                                        pnewname,
+                                                       CACHE_INODE_JOKER_POLICY,
                                                        &attrlookup,
                                                        ht,
                                                        pclient,
@@ -235,8 +236,8 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
                pentry_dirsrc, poldname->name, pentry_dirdest, pnewname->name);
 
       /* If the already existing object is a directory, source object should ne a directory */
-      if(pentry_lookup_dest->internal_md.type == DIR_BEGINNING &&
-         pentry_lookup_src->internal_md.type != DIR_BEGINNING)
+      if(pentry_lookup_dest->internal_md.type == DIRECTORY &&
+         pentry_lookup_src->internal_md.type != DIRECTORY)
         {
           V_w(&pentry_dirsrc->lock);
           if(pentry_dirsrc != pentry_dirdest)
@@ -251,8 +252,8 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
           return *pstatus;
         }
 
-      if(pentry_lookup_dest->internal_md.type != DIR_BEGINNING &&
-         pentry_lookup_src->internal_md.type == DIR_BEGINNING)
+      if(pentry_lookup_dest->internal_md.type != DIRECTORY &&
+         pentry_lookup_src->internal_md.type == DIRECTORY)
         {
           /* Return ENOTDIR */
           pclient->stat.func_stats.nb_err_unrecover[CACHE_INODE_RENAME] += 1;
@@ -291,7 +292,7 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
         }
 
       /* Entry with the newname exists, if it is a non-empty directory, operation cannot be performed */
-      if((pentry_lookup_dest->internal_md.type == DIR_BEGINNING) &&
+      if((pentry_lookup_dest->internal_md.type == DIRECTORY) &&
          (cache_inode_is_dir_empty(pentry_lookup_dest) != CACHE_INODE_SUCCESS))
         {
           /* The entry is a non-empty directory */
@@ -350,20 +351,10 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
 
   /* Get the handle for the dirsrc pentry */
 
-  if(pentry_dirsrc->internal_md.type == DIR_BEGINNING)
+  if(pentry_dirsrc->internal_md.type == DIRECTORY)
     {
-      phandle_dirsrc = &pentry_dirsrc->object.dir_begin.handle;
-      pattrsrc = &pentry_dirsrc->object.dir_begin.attributes;
-    }
-  else if(pentry_dirsrc->internal_md.type == DIR_CONTINUE)
-    {
-      P_r(&pentry_dirsrc->object.dir_cont.pdir_begin->lock);
-
-      phandle_dirsrc =
-          &pentry_dirsrc->object.dir_cont.pdir_begin->object.dir_begin.handle;
-      pattrsrc = &pentry_dirsrc->object.dir_cont.pdir_begin->object.dir_begin.attributes;
-
-      V_r(&pentry_dirsrc->object.dir_cont.pdir_begin->lock);
+      phandle_dirsrc = &pentry_dirsrc->object.dir.handle;
+      pattrsrc = &pentry_dirsrc->object.dir.attributes;
     }
   else
     {
@@ -381,21 +372,10 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
 
   /* Get the handle for the dirdest pentry */
 
-  if(pentry_dirdest->internal_md.type == DIR_BEGINNING)
+  if(pentry_dirdest->internal_md.type == DIRECTORY)
     {
-      phandle_dirdest = &pentry_dirdest->object.dir_begin.handle;
-      pattrdest = &pentry_dirdest->object.dir_begin.attributes;
-    }
-  else if(pentry_dirdest->internal_md.type == DIR_CONTINUE)
-    {
-      P_r(&pentry_dirdest->object.dir_cont.pdir_begin->lock);
-
-      phandle_dirdest =
-          &pentry_dirdest->object.dir_cont.pdir_begin->object.dir_begin.handle;
-      pattrdest =
-          &pentry_dirdest->object.dir_cont.pdir_begin->object.dir_begin.attributes;
-
-      V_r(&pentry_dirdest->object.dir_cont.pdir_begin->lock);
+      phandle_dirdest = &pentry_dirdest->object.dir.handle;
+      pattrdest = &pentry_dirdest->object.dir.attributes;
     }
   else
     {
@@ -448,8 +428,8 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
           fsal_status_t getattr_status;
 
           LogEvent(COMPONENT_CACHE_INODE,
-                   "cache_inode_rename: Stale FSAL File Handle detected for at least one in  pentry = %p and pentry = %p",
-                   pentry_dirsrc, pentry_dirdest);
+                   "cache_inode_rename: Stale FSAL File Handle detected for at least one in  pentry = %p and pentry = %p, fsaL_status=(%u,%u)",
+                   pentry_dirsrc, pentry_dirdest, fsal_status.major, fsal_status.minor);
 
           /* Use FSAL_getattrs to find which entry is staled */
           getattr_status = FSAL_getattrs(phandle_dirsrc, pcontext, &attrlookup);
@@ -459,7 +439,7 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
                        "cache_inode_rename: Stale FSAL File Handle detected for pentry = %p",
                        pentry_dirsrc);
 
-              if(cache_inode_kill_entry(pentry_dirsrc, ht, pclient, &kill_status) !=
+              if(cache_inode_kill_entry(pentry_dirsrc, WT_LOCK, ht, pclient, &kill_status) !=
                  CACHE_INODE_SUCCESS)
                 LogCrit(COMPONENT_CACHE_INODE,
                         "cache_inode_rename: Could not kill entry %p, status = %u",
@@ -467,17 +447,18 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
             }
 
           getattr_status = FSAL_getattrs(phandle_dirdest, pcontext, &attrlookup);
-          if(getattr_status.major == ERR_FSAL_ACCESS)
+          if( ( getattr_status.major == ERR_FSAL_ACCESS ) && ( pentry_dirdest != pentry_dirsrc ) )
             {
               LogEvent(COMPONENT_CACHE_INODE,
                        "cache_inode_rename: Stale FSAL File Handle detected for pentry = %p",
                        pentry_dirdest);
 
-              if(cache_inode_kill_entry(pentry_dirdest, ht, pclient, &kill_status) !=
+              if(cache_inode_kill_entry(pentry_dirdest, WT_LOCK, ht, pclient, &kill_status) !=
                  CACHE_INODE_SUCCESS)
                 LogCrit(COMPONENT_CACHE_INODE,
                         "cache_inode_rename: Could not kill entry %p, status = %u",
                         pentry_dirdest, kill_status);
+
             }
 
           *pstatus = CACHE_INODE_FSAL_ESTALE;
@@ -534,7 +515,9 @@ cache_inode_status_t cache_inode_rename(cache_entry_t * pentry_dirsrc,
       status = cache_inode_add_cached_dirent(pentry_dirdest,
                                              pnewname,
                                              pentry_lookup_src,
-                                             NULL, ht, pclient, pcontext, pstatus);
+                                             ht, 
+					     &new_dir_entry,
+					     pclient, pcontext, pstatus);
       if(status != CACHE_INODE_SUCCESS)
         {
           pclient->stat.func_stats.nb_err_unrecover[CACHE_INODE_RENAME] += 1;
